refactor(server): split match.ts — extract spectate + replay handlers (S27.8.16)
First slice on `apps/server/src/routes/match.ts` (was 1766 lines).
Extract the two simplest, most isolated read-only handlers
(`handleSpectateMatch` and `handleReplayMatch`) into a dedicated
module `routes/match-spectate-replay-handlers.ts` (189 lines).

Both handlers are thematically cohesive (read-only state/turns
exposure for spectators and replay viewers) and have no shared state
with the lifecycle handlers (create/join/accept/move) that remain
in `match.ts`.

Re-export from `match.ts` to preserve the public API consumed by
integration tests and any external module that imported these
handlers directly.

Dependencies are leaf-only: `prisma`, `getSpectatorCount`,
`serverLog`, `sendError/sendSuccess`, `AuthenticatedRequest`. No
cycle back to `match.ts`.

`match.ts` shrinks 1766 -> 1631 lines (-135 net lines). Several more
slices will follow to drop the file under a reasonable target.

- 1787/1787 server tests pass (no regression)
- typecheck clean

`team.ts` (2414l) untouched in this slice.
